Created cli file and moved index content to cli
